+++ src/deluge/gui/menu_item/menu_item.h
@@ -244,5 +244,12 @@ public:
 	/// Internal helper which can draw the standard deluge::gui::menu_item::Submenu layout.
 	static void drawItemsForOled(std::span<std::string_view> options, int32_t selectedOption, int32_t offset = 0);
 
+	/// @brief Check if selecting this menu item (with select encoder) should enter a submenu
+	virtual bool shouldEnterSubmenu() { return true; }
+
+	/// @brief Handle rendering of submenu item types
+	virtual int32_t getSubmenuItemTypeRenderLength() { return (3 + (kTextSpacingX * 3) + kTextSpacingX); }
+	virtual void renderSubmenuItemTypeForOled(int32_t xPixel, int32_t yPixel);
+
 	/// @}
 };