+++ src/deluge/gui/menu_item/patched_param/integer.h
@@ -57,6 +57,14 @@ public:
 		MenuItemWithCCLearning::learnKnob(fromDevice, whichKnob, modKnobMode, midiChannel);
 	};
 
+	// renders param value in submenus after the item name
+	void renderSubmenuItemTypeForOled(int32_t xPixel, int32_t yPixel) override;
+
+	int32_t getParamValue() {
+		readCurrentValue();
+		return getValue();
+	}
+
 protected:
 	void readCurrentValue() override;
 	void writeCurrentValue() final;