+++ src/deluge/gui/ui/audio_recorder.cpp
@@ -129,34 +129,30 @@ void AudioRecorder::renderOLED(deluge::hid::display::oled_canvas::Canvas& canvas
 }
 
 bool AudioRecorder::setupRecordingToFile(AudioInputChannel newMode, int32_t newNumChannels,
-                                         AudioRecordingFolder folderID) {
+                                         AudioRecordingFolder folderID, bool writeLoopPoints, bool shouldNormalize) {
 
 	if (ALPHA_OR_BETA_VERSION && recordingSource > AudioInputChannel::NONE) {
 		FREEZE_WITH_ERROR("E242");
 	}
 
-	recorder =
-	    AudioEngine::getNewRecorder(newNumChannels, folderID, newMode, false, false, kInternalButtonPressLatency);
+	recorder = AudioEngine::getNewRecorder(newNumChannels, folderID, newMode, false, writeLoopPoints,
+	                                       kInternalButtonPressLatency);
 	if (!recorder) {
 		display->displayError(Error::INSUFFICIENT_RAM);
 		return false;
 	}
 
 	recorder->allowFileAlterationAfter = true;
+	recorder->allowNormalization = shouldNormalize;
 
 	recordingSource = newMode; // This sets recording to begin happening even as the file is created, below
 
 	return true;
 }
 
-bool AudioRecorder::beginOutputRecording() {
-	AudioRecordingFolder folder = AudioRecordingFolder::RESAMPLE;
-	AudioInputChannel channel = AudioInputChannel::OUTPUT;
-	if (stemExport.processStarted) {
-		folder = AudioRecordingFolder::STEMS;
-		channel = AudioInputChannel::MIX;
-	}
-	bool success = setupRecordingToFile(channel, 2, folder);
+bool AudioRecorder::beginOutputRecording(AudioRecordingFolder folder, AudioInputChannel channel, bool writeLoopPoints,
+                                         bool shouldNormalize) {
+	bool success = setupRecordingToFile(channel, 2, folder, writeLoopPoints, shouldNormalize);
 
 	if (success) {
 		indicator_leds::blinkLed(IndicatorLED::RECORD, 255, 1);