+++ src/deluge/gui/ui/audio_recorder.h
@@ -34,7 +34,9 @@ public:
 	bool getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows);
 
 	ActionResult buttonAction(deluge::hid::Button b, bool on, bool inCardRoutine);
-	bool beginOutputRecording();
+	bool beginOutputRecording(AudioRecordingFolder folder = AudioRecordingFolder::RESAMPLE,
+	                          AudioInputChannel channel = AudioInputChannel::OUTPUT, bool writeLoopPoints = false,
+	                          bool shouldNormalize = true);
 	void process();
 	void slowRoutine();
 	bool isCurrentlyResampling();
@@ -53,7 +55,8 @@ public:
 
 private:
 	void finishRecording();
-	bool setupRecordingToFile(AudioInputChannel newMode, int32_t newNumChannels, AudioRecordingFolder folderID);
+	bool setupRecordingToFile(AudioInputChannel newMode, int32_t newNumChannels, AudioRecordingFolder folderID,
+	                          bool writeLoopPoints = false, bool shouldNormalize = true);
 };
 
 extern AudioRecorder audioRecorder;