+++ src/deluge/model/settings/runtime_feature_settings.h
@@ -61,6 +61,7 @@ enum RuntimeFeatureSettingType : uint32_t {
 	EnableDxShortcuts,
 	EmulatedDisplay,
 	EnableKeyboardViewSidebarMenuExit,
+	EnableLaunchEventPlayhead,
 	MaxElement // Keep as boundary
 };
 
@@ -90,6 +91,7 @@ public:
 
 	// Traded type safety for option values for code simplicity and size, use enum from above to compare
 	inline uint32_t get(RuntimeFeatureSettingType type) { return settings[type].value; };
+	inline bool isOn(RuntimeFeatureSettingType type) { return get(type) == RuntimeFeatureStateToggle::On; }
 
 	/**
 	 * Set a runtime feature setting.